refactor(cli): extract engineering commands to cli_code.py
First safe cut at the cli.py god-file (3415 LOC): move the cohesive,
test-covered engineering command handlers (_cmd_workspace/_cmd_task/
_cmd_gitops/_cmd_codereview/_cmd_patch/_cmd_code, ~374 lines) into cli_code.py.
They were already self-contained (args in, exit code out, lazy module imports),
so the move is behavior-preserving; build_parser imports them for set_defaults.
Full suite green. Remaining command groups can follow the same pattern
incrementally.
